1) split StockService into serveral sub functions; 2) extract folder path setting into property file and refactor excelHelpers; 3) add new Query function for new roleName; 4) record inStock-type during inStock process; 5) functions for user management(on-going); 6) code clean-up
