+++ wmstool/src/main/java/com/wmstool/wmstool/models/RoleName.java
@@ -1,5 +1,5 @@
 package com.wmstool.wmstool.models;
 
 public enum RoleName {
-	ROLE_Normal, ROLE_Operator, ROLE_Admin
+	ROLE_Normal, ROLE_Sales, ROLE_Operator, ROLE_Admin
 }
