+++ wmstool/src/main/java/com/wmstool/wmstool/services/HistoryService.java
@@ -5,7 +5,6 @@ import java.util.List;
 import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 
 import com.wmstool.wmstool.models.History;
@@ -14,7 +13,6 @@ import com.wmstool.wmstool.repositories.StockIdentifierRepo;
 import com.wmstool.wmstool.utilities.HistoryTreeNode;
 
 @Service
-//@PreAuthorize("hasRole('Role_Operator') or hasRole('Role_Admin')")
 public class HistoryService {
 
 	@Autowired
