+++ wmstool/src/main/java/com/wmstool/wmstool/utilities/DailyStockComparisonExcelHelper.java
@@ -18,18 +18,17 @@ import org.apache.poi.ss.usermodel.Sheet;
 import org.apache.poi.ss.usermodel.VerticalAlignment;
 import org.apache.poi.ss.usermodel.Workbook;
 import org.apache.poi.ss.usermodel.WorkbookFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.core.env.Environment;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
 
+@Component
 public class DailyStockComparisonExcelHelper {
 
-	@Autowired
-	private static Environment env;
+	@Value("${file.filePathForExcels}")
+	private String folderPath;
 
 	private static final String seperator = File.separator;
-	private static final String parentDir = env.getProperty("filePathForExcels") + seperator + "DailyStockComparison";
 	private static final String filetype = ".xls";
-	private static final String templateFile = parentDir + seperator + "DailyStockComparisonTemplate" + filetype;
 	private static final String filenamePrefix = "DailyStockComparison-";
 	private static final DateTimeFormatter dtf = DateTimeFormatter.ofPattern("yyyyMMdd");
 
@@ -37,7 +36,10 @@ public class DailyStockComparisonExcelHelper {
 	 * Create a excel from template excel which is named by current date formated in
 	 * "yyyyMMdd" pattern
 	 */
-	public static String createNewFile(LocalDate now) throws IOException {
+	public String createNewFile(LocalDate now) throws IOException {
+		String parentDir = folderPath + seperator + "DailyStockComparison";
+		String templateFile = parentDir + seperator + "DailyStockComparisonTemplate" + filetype;
+
 		// Read template file
 		Workbook workbook = WorkbookFactory.create(new File(templateFile));
 
@@ -68,8 +70,10 @@ public class DailyStockComparisonExcelHelper {
 	/**
 	 * Using given compareResult information to update the given filename excel
 	 */
-	public static void outputComparisonResult(List<List<String>> compareResult, LocalDate now, String fileName)
+	public void outputComparisonResult(List<List<String>> compareResult, LocalDate now, String fileName)
 			throws IOException {
+		String parentDir = folderPath + seperator + "DailyStockComparison";
+
 		// Read target file
 		String fileFullName = parentDir + seperator + now.getYear() + seperator + now.getMonthValue() + seperator
 				+ fileName;
